refactor(core): fold drag-distance/rectangle wrappers into dragHandlers
The calculateDragDistance/Rectangle 1-line wrappers were thin closures
over store.dragStart that only the drag-handlers controller needs.
Move them inside the controller so init() doesn't need the wrappers,
the utility imports, or the store destructure.

index.tsx: 688 -> 679 (-9). 15 drag-selection specs green.
